// frontend/src/features/products/types.ts
export type Product = {
  id: number;
  name: string;
  gvtId: number;
  productTagline: string;
  shortDescription: string;
  longDescription: string;
  logoLocation: string;
  productUrl: string;
  voucherTypeName: string;
  orderUrl: string;
  productTitle: string;
  variableDenomPriceMinAmount?: string;
  variableDenomPriceMaxAmount?: string;
  __typename: string;
};